# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Observer API",
        "version": "0.1"
    },
    "paths": {
        "/api/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Find list of Posts",
                "operationId": "App\\Http\\Controllers\\PostController::find",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category of Posts ('politics', 'economics', 'health', 'sport', 'culture', 'society')",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of Posts ('news', 'article', 'blog post', 'quote')",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "flag",
                        "in": "query",
                        "description": "Flag attributed to Posts ('main', 'popular')",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/elems"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found Posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Save Post",
                "operationId": "App\\Http\\Controllers\\PostController::save",
                "requestBody": {
                    "description": "Post data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Saved Post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Find Post by id",
                "operationId": "App\\Http\\Controllers\\PostController::findById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of the Post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found Post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "posts"
                ],
                "summary": "Delete Post",
                "operationId": "App\\Http\\Controllers\\PostController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of the Post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted"
                    }
                }
            }
        },
        "/api/posts/search": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Find list of Posts",
                "operationId": "App\\Http\\Controllers\\PostController::search",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Query to search Posts by",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/elems"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found Posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{id}/flag": {
            "put": {
                "tags": [
                    "posts"
                ],
                "summary": "Assign flag to Post",
                "operationId": "App\\Http\\Controllers\\PostController::flag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of the Post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "flag",
                        "in": "query",
                        "description": "Flag for the Post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "main",
                                "popular"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post flag assigned"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Find Categories",
                "operationId": "App\\Http\\Controllers\\PostController::findCategories",
                "responses": {
                    "200": {
                        "description": "Found Categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Author": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Post": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "subtitle": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "type": {
                        "type": "integer",
                        "enum": [
                            "News",
                            "Article",
                            "BlogPost",
                            "Quote"
                        ]
                    },
                    "author": {
                        "$ref": "#/components/schemas/Author"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "added": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "image": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "page": {
                "name": "page",
                "in": "query",
                "description": "Page number",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 1
                }
            },
            "elems": {
                "name": "elems",
                "in": "query",
                "description": "Number of elements per page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 10
                }
            }
        }
    }
}